refactor: update storageLocation filename
