+++ apps/desktop/src/app/accounts/settings.component.ts
@@ -11,7 +11,7 @@ import { PlatformUtilsService } from "@bitwarden/common/abstractions/platformUti
 import { StateService } from "@bitwarden/common/abstractions/state.service";
 import { VaultTimeoutSettingsService } from "@bitwarden/common/abstractions/vaultTimeout/vaultTimeoutSettings.service";
 import { DeviceType } from "@bitwarden/common/enums/device-type";
-import { StorageLocation } from "@bitwarden/common/enums/storageLocation";
+import { StorageLocation } from "@bitwarden/common/enums/storage-location";
 import { ThemeType } from "@bitwarden/common/enums/themeType";
 import { Utils } from "@bitwarden/common/misc/utils";
