+++ libs/common/src/services/state.service.ts
@@ -19,7 +19,7 @@ import { EnvironmentUrls } from "../auth/models/domain/environment-urls";
 import { KdfConfig } from "../auth/models/domain/kdf-config";
 import { HtmlStorageLocation } from "../enums/html-storage-location";
 import { KdfType } from "../enums/kdf-type";
-import { StorageLocation } from "../enums/storageLocation";
+import { StorageLocation } from "../enums/storage-location";
 import { ThemeType } from "../enums/themeType";
 import { UriMatchType } from "../enums/uriMatchType";
 import { StateFactory } from "../factories/stateFactory";
